feat: Renamed Annotation folder to Layer
